Fix inventory crop edge case
When the player location is touching the left of the stockpile box and also to the right of the center of the screenshot, the black outline of the location arrow could cause a box to be detected around that.

This change checks the left and right borders of the inventory box to make sure they're at least 80% dark pixels.  Since the middle of the location arrow is orange (not dark) this non-inventory box fails the test.